Fixed where the default value for status and auditedFlag were not being set.  This one took a while to track down.  The stuffTheProject function was overwriting the default values set in the model.
Also made it show the correct status of a failed create call.  It was returning a 200 with the error details.  Now it returns the proper response and shows the user what was missing.  Still needs some UI work to show all errors.  It only shows the first error it gets right now.